Refactored python code into package
