+++ .gitignore
@@ -17,3 +17,5 @@ test-bin/
 Makefile.paths
 SysInfo.chpl
 Untitled.ipynb
+dist/
+arkouda.egg-info/
