+++ VERSION
@@ -1,0 +1,1 @@
+2019.10.22
